Add screenshots for finding API keys
Added screenshots to show where to find API keys in Akismet, reCaptcha,
and Scribd. SendGrid does not require a screenshot, as we're asking the
user to enter their username and password.

These screenshots will have to be kept up to date as the underlying
services change. Also, I'm not sure that the process for signing up for
developer access on Scribd is clear enough and may require additional
screenshots.